+++ integration_tests/suite/base/test_timezones.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0+
+
+from hamcrest import assert_that, has_items, not_
+from . import confd
+
+
+def test_get_destinations():
+    response = confd.timezones.get()
+    response.assert_ok()
+
+    assert_that(response.items, has_items({'zone_name': 'America/Montreal'}))
+    assert_that(response.total, not_(0))